Added support for capturing party events (I think... not exhaustively tested yet).
Also created the function tl_parse_duration(), which does essentially the same thing as tl_parse_time() [renamed from parseTime()] but uses the time-to-complete rather than the finish time. This gives greater accuracy and robustness, as this is often given with second precision while the completion time is often listed to the minute.
This new function has not yet been exhaustively tested or completely implimented throughout the code, but is now used for parties, for building, and for market events.

Updated TODO again...