This adds mouseover and mouseout listeners to the events in the tooltip, such that the header's resource values get updated to reflect what their value would be when the event in question occurs.
This is done by a simple rework of the make_header() function to take it's base time as a parameter. On mouseover of an event, pass it the time of the event and you get the value the header would be at this time. Also had to rework what gets passed to the tooltip.add function, so that the times in 'contents' no longer get stripped off. This means we can no longer use 'join', but c'est la vie. This particular part is probably faster now, anyways (both the element stripper and the join itterate through 'contents', now only one itteration is needed).