This adds full merchant predictive powers!
The merchant collector now interpolates for 'from' and 'return' merchant events, based on the 'to' event and without resorting to language dependencies. 'from' events are distinguished from 'to' events by counting merchants, and always running over the merchant events from bottom to top.

I've done some moderate testing on this, but... it's pretty complex, so I'm sure I missed something.