Merchant prediction, take 2...
I don't think we can beat the language dependency here - may as well join it. :P The stupid 'sell' function was screwing with the merchant counts in a way we can't predict... so unless someone has a brainwave...

We'd have to use the language for creating the messages, anyways. And this way, we don't have to save anything extra in the events to distinguish between incoming and outgoing merchants!