This modifies the read(), write(), and remove() functions to allow accessing arbitrary scopes from arbitrary scopes.
Thus, if I'm in scope <server1>.<user1>, I can access data from scope <server2>.<user2>. This kind of functionality is needed by the unnatural timeline, and is needed to extend the functionality of the normal timeline to display events from other users simultaneously with local data. It also allows for a simpler way to merge settings from scope to scope (no need to export to global scope, than reimport).